# docs/swagger.yaml
info:
  contact: {}
paths:
  /api/cases:
    get:
      description: Retrieve all cases from the database
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Cases
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all cases
      tags:
      - Cases
  /api/cases/{id_case}/suspects:
    get:
      description: Retrieve all suspects for a specific case
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id_case
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Suspect Data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get all suspects
      tags:
      - Suspects
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new suspect for a specific case
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id_case
        required: true
        type: string
      - description: ID Card Number
        in: formData
        name: id_card_number
        required: true
        type: string
      - description: Full Name
        in: formData
        name: full_name
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Alibi
        in: formData
        name: alibi
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Success Create Suspect
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new suspect
      tags:
      - Suspects
  /api/cases/{id_case}/suspects/{id}:
    delete:
      description: Delete a suspect from a case
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id_case
        required: true
        type: string
      - description: Suspect ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Delete Suspect Data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete suspect
      tags:
      - Suspects
    get:
      description: Retrieve a specific suspect from a case
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id_case
        required: true
        type: string
      - description: Suspect ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Suspect Data
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get suspect by ID
      tags:
      - Suspects
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update an existing suspect information
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id_case
        required: true
        type: string
      - description: Suspect ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: ID Card Number
        in: formData
        name: id_card_number
        required: true
        type: string
      - description: Full Name
        in: formData
        name: full_name
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        required: true
        type: string
      - description: Alibi
        in: formData
        name: alibi
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Update Suspect Data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update suspect data
      tags:
      - Suspects
  /api/cases/{id}:
    get:
      description: Retrieve a specific case using its UUID
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Case By Id
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get case by ID
      tags:
      - Cases
  /api/cases/case-number/{number}:
    get:
      description: Retrieve a case using its case number
      parameters:
      - description: Case Number
        in: path
        name: number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Get Case By Case Number
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get case by case number
      tags:
      - Cases
  /api/cases/create:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a new case with title, description, incident date, and location
      parameters:
      - description: Case Title
        in: formData
        name: case_title
        required: true
        type: string
      - description: Case Description
        in: formData
        name: case_description
        required: true
        type: string
      - description: Incident Date (YYYY-MM-DD)
        in: formData
        name: incident_date
        required: true
        type: string
      - description: Case Location
        in: formData
        name: location
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Success Create Case
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new case
      tags:
      - Cases
  /api/cases/delete/{id}:
    delete:
      description: Delete a case by its ID
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Delete Case
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a case
      tags:
      - Cases
  /api/cases/update/{id}:
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update an existing case with new title, description, incident date,
        and location
      parameters:
      - description: Case ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Case Title
        in: formData
        name: case_title
        required: true
        type: string
      - description: Case Description
        in: formData
        name: case_description
        required: true
        type: string
      - description: Incident Date (YYYY-MM-DD)
        in: formData
        name: incident_date
        required: true
        type: string
      - description: Case Location
        in: formData
        name: location
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success Update Case
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update a case
      tags:
      - Cases
swagger: "2.0"